Add files via upload
This repository contains MATLAB codes for Content-Based Image Retrieval made with a GUI to be clear and easy to use. The images dataset which the search is made on are stored in a the folder “images”, the main GUI is coded in the two files “CBIR.fig” and”CBIR.m” but the process of features extraction is made by the code “Extract_features.m”. First the query image is loaded at this point all the previously mentioned features are extracted from the image then it is shown in the main GUI platform under title “Loaded image”. Then the extracted features are compares to the already saved and processed database where the distance between the query image and all images in the dataset is calculated. Finally the nearest ten images to the query image are shown in the GUI.